refactor: adopt package oriented design
