+++ Makefile
@@ -3,7 +3,7 @@ CIRCLE_SHA1 ?= latest
 default: run
 
 run:
-	go run main.go
+	go run cmd/server/main.go
 
 build:
 	mkdir -p ./dist
@@ -11,7 +11,7 @@ build:
 		go build -v \
 		-ldflags "-X main.CommitSha=${CIRCLE_SHA1}" \
 		-o ./dist/buran \
-		./
+		./cmd/server
 
 test:
 	go test -v ./...
