+++ internal/cache/cache_provider.go
@@ -1,12 +1,13 @@
 package cache
 
 import (
-	"github.com/escaleseo/buran/cache/memory"
-	"github.com/escaleseo/buran/cache/redis"
-	"github.com/escaleseo/buran/cache/rediscluster"
-	"github.com/escaleseo/buran/env"
 	"github.com/gregjones/httpcache"
 	"github.com/pkg/errors"
+
+	"github.com/escaleseo/buran/internal/cache/memory"
+	"github.com/escaleseo/buran/internal/cache/redis"
+	"github.com/escaleseo/buran/internal/cache/rediscluster"
+	"github.com/escaleseo/buran/internal/platform/env"
 )
 
 type Provider interface {
