+++ internal/platform/logger/middleware.go
